Addressed crash for opening intent outside of activity and library bump
Was getting a crash and needed to force a new task
Why does android care where we start Intents from
isn't that the point?
Anyway passed in the activity context instead of the
app context from the main page. Added the new task
flag for the deeplink intent and not sure that's really
required, but did it anyway as it doesn't have an activity
intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)